Cleanup: Move Substitute Day setting utils into it's module
