+++ ui/src/components/timetables/substitute-day-settings/CommonSubstitutePeriod/CommonSubstitutePeriodForm.tsx
@@ -10,17 +10,17 @@ import { useAppDispatch } from '../../../../hooks';
 import { setIsCommonSubstitutePeriodFormDirtyAction } from '../../../../redux/slices/timetable';
 import { mapToISODate, padToTwoDigits } from '../../../../time';
 import { DateRange } from '../../../../types';
-import {
-  mapDateTimeToFormState,
-  mapLineTypes,
-  parseSubstituteDayOfWeek,
-  submitFormByRef,
-} from '../../../../utils';
+import { submitFormByRef } from '../../../../utils';
 import { SimpleButton } from '../../../common/Buttons';
 import { Row } from '../../../common/LayoutComponents';
 import { ConfirmationDialog } from '../../../common/Modals';
 import { useDirtyFormBlockNavigation } from '../../../forms/common/NavigationBlocker';
 import { commonSubstituteDayData } from '../common_substitute_day_data';
+import {
+  mapDateTimeToFormState,
+  mapLineTypes,
+  parseSubstituteDayOfWeek,
+} from '../utils';
 import {
   CommonDayType,
   FormState,
